File output. Py3. Clean-Up. Closes #3, Closes #5.
