--- conf/.gitignore
@@ -1,3 +1,0 @@
-*.yaml
-!*.dist.yaml
-!logging.yaml
