--- conf/kako.dist.yaml
@@ -1,14 +1,0 @@
----
-logging:
-  path: '/opt/kako/logs'
-
-simulations:
-  path: '/opt/kako/simulations/'
-
-results:
-  region: 'us-west-2'
-  topic: ''
-
-monitoring:
-  enabled: True
-  interval: 60
