--- conf/routerlogin.pem
@@ -1,51 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQCxQeNhKw4FJNgE
-/+nifESwDpU+sS9yL5J4qTpRc6mcVTbGMjtvuaNF29Sohm++5iL9Po9j+xfbCImv
-JKC1Bvvk4V61x+Z6kpTFeEA6uujU0yitav45O+7PrIEVZ48AflTstoWxnU1M7VFo
-jUpz6IO0hmKKJQOz0W2kyZ5TkqXnP7YW7KjcLEIlVWp32Li8s0aVKDkBXkyjpqXg
-PO6DSMeLgYm8Mwsd8ks+JRLnnyL64XkVJiDTlB3vV0JUj+h/0BxhwecAh4uw54fw
-amUgYr8WKLihl1broFiRXVOfDF+Ohsy8nmIViNQxwzFej801jSR/T97oerTZEq3c
-/ecfrUxLAgMBAAECggEASOFwXglX0GGfuB6IyaiSjVjRoF81xNsufBIIemj1+eNH
-jCInMf6/1P9X5pqv42u4QyuvxUn+Ov69rZd/xC7E7wyC5nb0mr0eEjspHXywMqFH
-flYZhOrRe+3PfT+BGpDW1iUNv0OP/ussp7JARHv7ewQuC+LALPUVAcYsJOyM7hEi
-xHwfvZUl49QbkYpxgpC3zAF3hLjO3bNRmtrNamPBEYTuYuV6IzxuQ5Jv2D9rM4T+
-SoSbGmr6rmUfTkXhJ/glrS2/noaUGDFNYJxwrxS4GTFnqenLCxBtKaRdvV2hQI36
-DweYKYHaKCoNBTeR2kYySGz4brjstlV7IJz8liLokQKBgQDceNrBxcagGB0mcmNp
-3n7o9ha3iSRoVOjds309eWb+r48bwVvfvxS/N/awaSsmpgAYlcw7FpzSp5LN82X1
-p5/6yOJPzfvnuKKi+04v+FSYSHz66NEFrSulnRc28C5D5EtXQ8pGernnU09QI0mS
-7Bv7yHqUyvZ1O7fxSDE19PK8bQKBgQDN0kuMqii4jvo222PWEbumLd4c1rb/gxsy
-gKc3OyjL3kPeX0fpTCThn5t5jfvd5vBe2VmnbJwxaC9WfyyTf8OevnGa/rpbpRHS
-6ElI3XG0luVTK7xVMCJYwYbarSAha6aBJeSDal/swTopodX/cT+jIanm88Iyb17J
-+d1YMZfIlwKBgExePIA0brnTwH0p0E/CtzkgzPukMI1HJCZvXullXQzJaQvXLWjS
-enBGBVhUBZNsmsnpUYYCLi41x385WU7/UkX/QY0dL62zYfqMjCfHoB3jtg7/iC+s
-LVs/j+QCzDPOJ17B55YEOtLrI6UBpCGVUZgkeHmD2F0BY5TbSdiD0XT9AoGATBn9
-pT0lYVg8XBwyhex4nbFH9xHNCBzJ6Zp4OYu+QeMk5QwF4tUsrd/99Kvy/NFu/yMO
-a0FVX09OcYUETmVnR0LTbqPQ9CHMvVb2AeMfL867bE03jagGixrDagepYbIUFpyU
-KaMjZL7taodUQOb2yLKhx6tJmqxjN/sTRJDRtnECgYB/6rLD7K6vBGe/8nQApGAJ
-OFs3ACkWroZwUSZm+HkMDXoIl/FSs0mZTWWl/0/yipeVshcpQUTnapZUx8anhu7M
-sStpUOnidUbmwrlVPP+wsOenqqHKr49AqCs9/orkGb5EsKp0NcyBsiFFZrE+N0bc
-z95rQ08/oIWSKv04+IYqoA==
------END PRIVATE KEY-----
------BEGIN CERTIFICATE-----
-MIID4TCCAsmgAwIBAgIJAOvHdOVKK8rUMA0GCSqGSIb3DQEBCwUAMIGGMQswCQYD
-VQQGEwJVUzETMBEGA1UECAwKQ2FsaWZvcm5pYTERMA8GA1UEBwwIU2FuIEpvc2Ux
-EDAOBgNVBAoMB05FVEdFQVIxHzAdBgNVBAsMFkhvbWUgQ29uc3VtZXIgUHJvZHVj
-dHMxHDAaBgNVBAMME3d3dy5yb3V0ZXJsb2dpbi5uZXQwHhcNMTcwMTAzMDMzNzQ1
-WhcNMjcwMTAxMDMzNzQ1WjCBhjELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlm
-b3JuaWExETAPBgNVBAcMCFNhbiBKb3NlMRAwDgYDVQQKDAdORVRHRUFSMR8wHQYD
-VQQLDBZIb21lIENvbnN1bWVyIFByb2R1Y3RzMRwwGgYDVQQDDBN3d3cucm91dGVy
-bG9naW4ubmV0MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAsUHjYSsO
-BSTYBP/p4nxEsA6VPrEvci+SeKk6UXOpnFU2xjI7b7mjRdvUqIZvvuYi/T6PY/sX
-2wiJrySgtQb75OFetcfmepKUxXhAOrro1NMorWr+OTvuz6yBFWePAH5U7LaFsZ1N
-TO1RaI1Kc+iDtIZiiiUDs9FtpMmeU5Kl5z+2Fuyo3CxCJVVqd9i4vLNGlSg5AV5M
-o6al4Dzug0jHi4GJvDMLHfJLPiUS558i+uF5FSYg05Qd71dCVI/of9AcYcHnAIeL
-sOeH8GplIGK/Fii4oZdW66BYkV1TnwxfjobMvJ5iFYjUMcMxXo/NNY0kf0/e6Hq0
-2RKt3P3nH61MSwIDAQABo1AwTjAdBgNVHQ4EFgQUqKNrMYhA0PzF2NcYSen0v7Ep
-EPIwHwYDVR0jBBgwFoAUqKNrMYhA0PzF2NcYSen0v7EpEPIwDAYDVR0TBAUwAwEB
-/zANBgkqhkiG9w0BAQsFAAOCAQEArXdlmlt39oWNUeIhhqKLDK8pPnZaE00Dh9Kb
-koMeW0AQhe1kVR/qUPzhA/ngNNlGP1RCafJc5W/t5tiCYzY5wihjSQIB1Q5IOHBE
-6otd1ClGUWw6W6YaFd1MVdg0uYd36Je9ftoorG9Bi+JFdmEsuu5Kk+NGTp/6g26I
-nHAoJxBBMppQ2PzJneoKi46xxpQYFIQtjkpPHT9hCALpX0ow/PeVDAqGbBoyq5xR
-dmkQGWhVR1rzNIR2orYK7T8JFmvITgUy4CaxaPpOvovwU4jXpmC5p+u3WF2dFh9t
-VrmajRafvwDDLeugaHempJSRAcQKS1J2VcxpMFyK0U7gpFGq6A==
------END CERTIFICATE-----
