--- kako/__init__.py
@@ -1,8 +1,0 @@
-''' Kako - Manifest driven IoT honeypots. '''
-
-from kako import config
-from kako import constant
-from kako import messaging
-from kako import simulation
-
-__version__ = '0.2.0'
