--- kako/constant.py
@@ -1,3 +1,0 @@
-''' Implements configuration constants for Kako. '''
-
-DEFAULT_CONFIGURATION_FILE = 'conf/kako.yaml'
