--- requirements.txt
@@ -1,5 +1,0 @@
-click==6.6
-boto3==1.4.1
-PyYAML==3.12
-requests==2.11.1
-cerberus==1.0.1
