+++ sample/kako.yaml
@@ -1,0 +1,21 @@
+---
+logging:
+  path: './'
+
+simulations:
+  path: './sample/simulations/'
+
+results:
+  processor: file
+  attributes:
+    path: '/tmp/'
+
+# results:
+#   processor: sns
+#   attributes:
+#     region: 'us-west-2'
+#     topic: 'TOPIC_NAME'
+
+monitoring:
+  enabled: True
+  interval: 60
