--- simulations/000-LinuxTelnet.yaml
@@ -1,7 +1,0 @@
----
-name: Linux Generic Telnet Simulation
-port: 2323
-version: 0.1.0
-protocol: telnet
-server:
-  banner: 'Busybox Login'
