--- simulations/001-CPEServer.yaml
@@ -1,24 +1,0 @@
----
-name: CPEServer HTTP Simulation
-port: 7547
-version: 0.1.0
-protocol: http
-server:
-  banner: 'gSOAP/2.7'
-  response:
-    code: 404
-    text: 'Not Found'
-    body: ''
-    headers: []
-  routes:
-    get: []
-    options: []
-    head: []
-    post:
-      - vulnerability: 'Generic - TR-069 SOAP remote code execution'
-        route: '/UD/act'
-        response:
-          code: 200
-          text: 'OK'
-          body: ''
-          headers: []
