--- simulations/003-D-Link.yaml
@@ -1,49 +1,0 @@
----
-name: D-Link HTTP Simulation
-port: 8080
-version: 0.1.0
-protocol: http
-server:
-  banner: 'httpd'
-  response:
-    code: 401
-    text: 'Unauthorized'
-    body: ''
-    headers:
-      - key: 'WWW-Authenticate'
-        value: 'Basic realm="D-Link Router"'
-  routes:
-    options: []
-    head: []
-    post:
-      - vulnerability: 'D-Link - hedwig.cgi information disclosure'
-        route: '/hedwig.cgi'
-        response: 
-          code: 200
-          text: 'OK'
-          body: ''
-          headers:
-            - key: 'Content-Type'
-              value: 'text/xml'
-      - vulnerability: 'D-Link - command.php remote code execution'
-        route: '/command.php'
-        response:
-          code: 200
-          text: 'OK'
-          body: '610cker'
-          headers: []
-      - vulnerability: 'D-Link - HNAP1 multiple vulnerabilities'
-        route: '/HNAP1'
-        response:
-          code: 200
-          text: 'OK'
-          body: ''
-          headers: []
-    get:
-      - vulnerability: 'D-Link - Language remote code execution'
-        route: '/language/'
-        response:
-          code: 200
-          text: 'OK'
-          body: '610cker'
-          headers: []
