--- simulations/004-Rompager.yaml
@@ -1,19 +1,0 @@
----
-name: RomPager HTTP Simulation
-port: 5555
-version: 0.1.0
-protocol: http
-server:
-  banner: 'RomPager/4.07 UPnP/1.0'
-  response:
-    code: 200
-    text: 'OK'
-    body: ''
-    headers:
-      - key: 'EXT'
-        value: ''
-  routes:
-    options: []
-    head: []
-    post: []
-    get: []
