--- simulations/005-Huawei.yaml
@@ -1,27 +1,0 @@
----
-name: Huawei uPnP Simulation
-port: 37215
-version: 0.1.0
-protocol: http
-server:
-  banner: 'Linux uPnP/1.0 Huawei-ATP-IGD'
-  response:
-    code: 200
-    text: 'OK'
-    body: ''
-    headers:
-      - key: 'LOCATION'
-        value: 'http://192.168.1.1:37215/gatedesc.xml'
-      - key: 'CACHE-CONTROL'
-        value: 'max-age=86580'
-      - key: 'EXT'
-        value: ''
-      - key: 'ST'
-        value: 'upnp:rootdevice'
-      - key: 'USN'
-        value: 'uuid:e7ca24dc-d9de-4264-a134-0169ab3448ce::upnp:rootdevice'
-  routes:
-    options: []
-    head: []
-    post: []
-    get: []
