--- simulations/007-Generic-uPnP.yaml
@@ -1,21 +1,0 @@
----
-name: Generic uPnP Simulation
-port: 1900
-version: 0.1.0
-protocol: http
-server:
-  banner: 'miniupnpd/1.0 UPnP/1.0'
-  response:
-    code: 200
-    text: 'OK'
-    body: ''
-    headers:
-      - key: 'Location'
-        value: 'http://192.168.1.1:1900/device.xml'
-      - key: 'USN'
-        value: 'uuid:8308e25f-5147-46ef-92e1-9f2b412761e3::upnp:rootdevice'
-  routes:
-    options: []
-    head: []
-    post: []
-    get: []
