--- simulations/008-Linux-uPnP.yaml
@@ -1,19 +1,0 @@
----
-name: Linux uPnP Simulation
-port: 49152
-version: 0.1.0
-protocol: http
-server:
-  banner: 'Linux/3.0.8, UPnP/1.0, Portable SDK for UPnP devices/1.6.18'
-  response:
-    code: 404
-    text: 'Not Found'
-    body: ''
-    headers:
-      - key: 'CONTENT-TYPE'
-        value: 'text/html'
-  routes:
-    options: []
-    head: []
-    post: []
-    get: []
