+++ src/kako/processor/__init__.py
@@ -1,0 +1,4 @@
+''' Implements Kako results processors. '''
+
+from kako.processor import sns
+from kako.processor import file
