+++ tox.ini
@@ -1,0 +1,8 @@
+[tox]
+envlist = py36
+
+[testenv]
+deps=
+    pytest
+    pytest-cov
+commands=pytest --cov=kako
